lots of small stuff READ DESC
Cinemachine camera tweaks, camera should adjust faster now to large changes in position, and zoom in less when the players are close, meaning more of the level is visible on average, helping with visibility and reducing camera whiplash. Harpoon tips now break when the harpoon gets stuck in a surface. Harpoons kick up dust when they impact a surface. Harpoons now have a glint animation on the dangerous back tip. Players can now fast fall by pressing the down direction, accompanied by a particle effect showing the increased downward force. Players now have tunable physics materials, by default they have no friction now and a little bit of bounce. Players now have a heart that will pop out of their character when damaged (This was just me messing around having fun, this can change or be removed based on feedback). Changed the existing geyser bubble VFX that was in the level to be more in line with the art style of the player bubble VFX and clash VFX. Lowered the opacity of the score/HP banners as a short term fix for visibility near the top of the screen.