we got freaky teleporters
- added a teleporter prefab called HVTeleporter
- Allows for teleportation from a horizontal axis to a vertical axis and vice versa
- Drag Prefab to desired entrance location
- Line Up the Raycast with where you want the entrance to be
- Move the child object named "Connector" to the desired exit location
- Ensure the pivot of both the base prefab and the connectors, are lined up on the respective axis
- Ex. HVPrefab is placed on the top of the map acting as a teleporter if you go through it upwards. HVPrefab(1) is placed on the left of the map acting as a teleporter if you go through it to the left. HVPrefab should be placed on the same World X position as HVPrefab(1)'s child object "Connector". And HVPrefab(1) should be placed on the same World Y position as HVPrefab's child object "Connector".
- On each instance of the prefab, find the dropdown menu to select one of 4 ejection directions. Directions are as follows: Left, Right, Up, Down.
- Using the above example entering through HVPrefab would mean that the player emerges from the left side of the map with a velocity in the ---> direction
- So on the HVPrefab object, in the inspector, you need to click "right" on the drop down menu.
- And in this example HVPrefab(1)'s entrance would be on the left side of the map, where the player would emerge from the top side of the map
- So for HVPrefab(1), since we want the player to emerge down from the top, we would assign "Down" in the inspetor for that object.

How it works:

The prefab has a serialized bool called Is Vertical. This should be checked as true in the inspector if the portals intended location is either at the top of the map or the bottom.

If Is Vertical is true then the raycast firest horizontally and gets the players location on the x axis.
If Is Vertical is false then the raycast fires vertically and gets the players location on the y axis.

The raycast is represented by a Red line in the editor view to help with visualization.

If is Vertical is true and the player collides with the raycast, the raycast detects how far that player is from the Raycast pivot point on the X-axis and pops the player out along the Y-Axis of the child object "connector" at an equal magnitude.

If is Vertical is false and the player collides with the raycast, the raycast detects how far that player is from the Raycast pivot point on the Y-axis and pops the player out along the X-Axis of the child object "connector" at an equal magnitude.

Weapons work exactly the same when entering and exiting the portals.

However there is a serialized int that is currently set as = 1

This represents how many times an instance of a weapon can be teleported. This prevents the situation from the Portal Games where a player with a portal directly above them and below them can fall forever.